Moving Channels package to anticipate refactoring for rxjs
