+++ function/src/cli/index.ts
@@ -1,3 +1,3 @@
-export * from "./channels/index.js"
+export * from "../channels/index.js"
 export * from "./main/di/index.js"
 export * from "./app/index.js"
