+++ function/src/cli/main/di/Configuration.ts
@@ -2,7 +2,7 @@ import {
    RandomArtTaskCall,
    RandomArtTaskReply,
 } from "../../../painting/message/index.js"
-import { ChannelWrapper } from "../../channels/ChannelWrapper.js"
+import { ChannelWrapper } from "../../../channels/ChannelWrapper.js"
 import { IRandomArtTaskEngine } from "../../../painting/index.js"
 import { IpldRegionMapRepository } from "../../../plotting/ipld/components/IpldRegionMapRepository.js"
